cp Bootstrap & TinyMCE  w/ rails assets:precompile
This commit moves the logic for copying Bootstrap glyphicons and TinyMCE skins to the public directory from the assets.rb initializer to a dedicated Rake task. The new task, `assets:copy_assets`, is invoked as part of the `assets:precompile` process. This change ensures that the file copying logic is only executed during the assets precompilation process, which prevents redundant copying of these files.